Still debugging the cart tester. Reads don't appear to be working. Addresses are sending fine.

+++ sw/cart_tester/cart_tester.h
@@ -7,5 +7,5 @@
 #pragma once
 
 void send_address(uint32_t address);
-uint32_t start_read();
-void verify_data(uint32_t data, uint32_t address);
+uint16_t start_read();
+void verify_data(uint16_t data, uint32_t address);
